API Restructure: Well positioning.
The more nested style looks a bit overengineered at the moment,
but it'll come in handy once we want to traverse through the
robot components.

We'll need to get wells and manipulate them at some point, so
putting off the proper structure doesn't really give us any gains
to productivity.