Major calibration data structure changes.
Supporting custom calibrations for particular wells and the
optional specification of particular instruments.

NOTE: Most of this is going to be rewritten, it's just that we
don't have a concept of an entire robot right now that contains
instruments in order to attach the calibration data to the
instruments rather than the containers.