pFusX: Protocol output ordered keys.
Still debating where the proper balance is here. On one hand, it's
good to make this as understandable to nontechnical people as
possible. On the other, the code can get mighty obscure and that
could also present bugs.

The heart of the issue is whether people should be expected to
work with {"foo: true, "bar": false}, {"bar": false, "foo": true}
interchangably.