Protocol: Removed obsolete documentation.
I was going to just use *args and **kwargs to store the action
queue, but then I realized the expansion and normalization will
have to happen for basically every task.

Opted for a data structure that, while inefficient, makes it easy
to work with for all the other handlers.

Eventually, I'd like to circle back here and make this expanded
data structure a lazy thing and have the methods only run as each
instruction is actually passed to the other systems.

It'll be a lot easier when I've written the other systems to
understand their ideal data formats...