Protocol: Motor Handler now does transfers.
Had to do a weird closure thing, but it's much cleaner than any
of the alternatives.

The values haven't been tested themselves; the point here is to
ensure that each of the things that should happen gets kicked off
when necessary.

Later, we'll move into each of these individual operations to
fully tweak their behaviors.

We'll need to add a few more calibration points as well,
particularly on the well Z axes, and of course right now I'm
not calibrating by axis.