Motor Control: Mode to wait for each command to complete.
We're setting this for default by now, it turns debug mode on
with a custom SmoothieBoard firmware command (M62), as specified
in drivers.motor.OpenTrons, and then keeps reading lines from
serial until that line ({'stat':0}) is present.

This only happens in cases where the CNCDriver subclass has a
DEBUG_ON command.