Protocol: run with generator that yields total progress.
Need to figure out a clean way of yielding the progress of all
the commands as well as the current command on a particular
handler (for example, if the transfer command requires a bunch
of G-Code instructions in the MotorHandler).

I think the best way to do this would be to have each Handler
tally up its total number of commands, return that information,
and then have its own iterator.

Then Protocol can just count all the dependent instructions and
keep track of the entire total.

Of course, this will leave us with a jumpy progress indication,
since virtualized robot instances are obviously going to go much
faster than the actual robot, but each instruction counts the
same in the overall total.