Protocol: Virtualized context much more robust.
We now run commands in the virtualized context when they're added
and also reinitialize the context when beginning a protocol run.

This allows the user to get much more immediate feedback on
whether or not a protocol can run as specified, both during the
definition (or import) phase and the run (or export) phase.

Cleaned up some tests that got some useful feedback from the
context; transfers were a little loose.

Also made all the ContextHandler commands that aren't transfer
just silently pass through with *args and **kwargs.

Later, we'll flesh out these commands, but we're focused on
transfer at the moment.