Calibration: Major responsibility overhaul (Deck -> Context).
Basically, the Virtual Robot deck can't hold calibration for a
number of reasons.

From a technical standpoint, the context is reinitialized and
that will cause the deck to require reinitialization, which can
really mess with the state and is tricky to ensure.

From a design standpoint, the responsibility of the Virtual Robot
is to keep track of the effects of a protocol on the deck layout
and inventory.

This will make things a lot simpler going forward because now we
have a clear distinction between operational data, such as
calibration, and portable data, such as well volumes, which would
be the same for any machine or operational context at a given
step within a given protocol.

There might still be some ugly souvenirs from the old design, but
hopefully none that will cause a great deal of bugs.

The one dangerous side effect of this change is that Session is
no longer testing for separate environments, but that whole
Session design was a stopgap to the ContextHandler, so it's also
not a big deal.

I just need to refactor later to remove Session entirely, since
Protocol (and the resulting ContextHandler) is the new way of
doing that and the Motor Controller is no longer a distinct
module.