Remove logging into intercom w/fake user id
This commit removes logging into intercom with a fake user ID.

Instead we turn the intercom app on with without a user id which enables
intercoms guest mode. If a user then chooses to log in, their chat
conversation will be persisted on the same chat window. If we instead
logged in the user into intercom with a fake user ID and the user
decides to log in, all of their chat history in the current intercom
window will be lost as intercom will switch user accounts and bring up a
new chat history..