Lowercase pipette axis
If you do a find all for axis.upper() and axis.lower(), you'll see that
Robot uppercases axis then smoothie both uppercases and lowercases it at
different times. Rabbithole.

This commit is to fix a bug in app 2.x, where if you load a protocol
with pipette axis='A', you cannot jog the plunger or move it at all.