Enable probing with both pipettes
Adjust offsets and probe position / dimensions to enable
probing for both pipettes. Also adds mounts to pipettes (this is a hack)
to enable offset values to be pulled from instruments so they can be used
in calibration. This update of the global dicts should also allow calibration
to persist between probing and running.