refactor(protocol-designer): Split up step edit form logic into types (#1707)
The step edit form contained a lot of type-specific form logic that could be abstracted away. In
order to accomplish an ideal generalized solution for step form validation, the component needed to
be reorganized. Furthermore, instead of acting as a single connected parent to far downstream
components that might want data from redux. This adds connected leaf components that partition those
props to the field that needs them, and away from the general clutter in the Step Edit Form.