fix(protocol-designer): Only show deck setup prompt text when selected
Previously the explanatory copy above the deck setup step was shown if deck setup was selected or
hovered. Now it is only shown when selected.

BREAKING CHANGE: When deck setup step is selected, hovering over other steps will not change the
appearance of the liquid state on the deck map.