refactor(discovery-client): Switch out bonjour module with mdns-js
`bonjour` was having trouble picking up mdns advertisements on multiple interfaces, which means it
couldn't handle the new link-local IPv4 setup for wired robots. `mdns-js` appears to play nicely
with multiple advertisements for the same service coming in over multiple interfaces (e.g. wifi and
usb-to-ethernet), which is why this switch was made.