fix(*): Change Makefile python env var to be better named, document
On windows it’s much harder to do per-invocation environment variables and the
common practice is to set environment variables in the system configuration. To
support this, the environment variable used for the Python executable for ‘make
install’ is now called OT_PYTHON to make it more clear what it is used for.

The environment variable is now also documented in contributing.md.