fix(update-server): Set the cwd to the venv when running an otupdate selftest
At least on a mac, on zsh, (this issue may be different depending on system and shell because of the
use of shell=True when invoking the server to selftest) the update server self test picks up the
code in the source directory rather than the module to self test. This breaks the unit tests. By
explicitly setting the cwd of the subshell used for the server under test to the root of the venv,
it all works.