feat(api): calibrate modules on run, poll driver thread (#2071)
At the beginning of a run, after homing the z axes, calibrate all compatible modules. Create a layer
of multithreading with redundancy protection when querying the tempdeck for it's current
temperature. Because this prevents exceptions when the buffer gets backed up, the run app can now
poll GET /modules/{serial}/data more frequently.

Closes #2034