feat(compute,api,update): Add sys ver to health, allow resin pull
This commit adds an environment variable in the 3.3.0 dockerfile, pulled from
package.json in the api, that is the version of the git repo that the system was
built from. This is sent in the /health endpoints of both the api server and
update server:

GET /health
200 OK
{
    "name": "opentrons-moon-moon",
    "api_version": "3.3.0-beta.1",
    "fw_version": "edge-6168d32",
    "logs": [
        "/logs/serial.log",
        "/logs/api.log"
    ],
    "system_version": "3.3.0-beta.1"
}

GET /server/update/health
200 OK
{
    "name": "opentrons-moon-moon",
    "updateServerVersion": "3.3.0-beta.1",
    "apiServerVersion": "3.3.0-beta.1",
    "smoothieVersion": "edge-6168d32",
    "systemVersion": "3.3.0-beta.1"
}

In addition, this commit removes the resin-updates.lock file that prevents the host
OS from pulling container changes.