fix(update-server): Fix issues with 3.2 api on 3.3 system (#2097)
This fixes issues with updating to a 3.2 api on a 3.3 system, and also issues
with running 3.2 on a 3.3 system.

API server wheels from before 3.3.0 do not have resource subdirectories and should not be
provisioned. The update server now inspects the version (from the filename) of the api wheel it is
installing and will not provision wheels whose version is prior to 3.3.0.

In addition, once the 3.2 api is uploaded, it will not have established the
proper setup scripts. This commit changes the container to properly fall back on
the scripts established in /usr/local/lib.