feat(protocol-designer): default form fields from old protocols (#2162)
As PD updates there is a chance that step form fields will change from version to version. The
defaulting logic added here will allow users to open a protocol that was made in an older version of Protocol Designer that may not have had certain fields defined. This way they will be defaulted
logically and avoid broken behavior.