feat(api): Ensure the wifi mac address stays fixed
We've had sporadic user reports of the wifi mac address changing, which defeats mac address
whitelisting. It does not appear that NetworkManager is currently set up to do this - there does not
seem to be a global config and the defaults should keep it the same - but just to be sure, we can
tell NetworkManager explicitly to use the hardware address when connecting to wifi. Note that the
MAC address will still be randomized when scanning, but this shouldn't be a problem.