fix(api): Do not bind the api server to localhost if socket specd
Previously, the API server would always bind to a hostname and port even if a socket was specified -
aiohttp does not consider these mutually exclusive, and the arguments in argparse default to
localhost:8080. This can cause conflicts when the API server tries to start on a robot, in an
unpredictable way. We now consider listening to a socket and listening on a hostname mutually
exclusive.

2256