feat(api): Add security type of each ssid to GET /wifi/list
We now add nmcli's idea of what the security type of each scanned SSID is to the
list returned by GET /wifi/list. That list now looks like this:

{
    "list": [
        {
            "ssid": "--",
            "signal": 100,
            "active": false,
            "security": "WPA2"
        },
        {
            "ssid": "linksys",
            "signal": 100,
            "active": false,
            "security": "WPA2 802.1X"
        },
    ]
}

The string indicates 802.1x if it contains "802.1X", WPA2 if it only has WPA2.
If there is no security the value will be the empty string.

Note that we do not support WEP or WPA1.

Closes #2284