feat(api): Add hardware_control submodule
The hardware_control submodule is the controller of the robot's hardware. Its API can move the robot
at a level of abstraction that portrays the robot as a whole, not including the labware. In this PR,
the module is a stub, and not used by default. Further PRs will introduce more functionality.

Closes #2232