feat(api): Flash the smoothie on api boot if versions don't match
We currently rely on the client to update the smoothie board. We shouldn't do that objectively - the
firmware is part of the robot system that the api is meant to controk, so the api should be in
charge - but also we need to build it into the api to work around a bug in our windows builders that
means installing via the client will not work.