feat(api): Add new protocol API stubs
The protocol API will be the way for protocols to interact with the OT2 system.
It will allow users to run protocols locally and simulate them as well as
allowing the higher level system to run the protocols on a robot. It will couple
to the hardware controller.

This commit introduces the structure of the API. It is conceptually similar to
the robot API we currently use, but explicitly notes that what you manipulate
are contexts rather than exact hardware representations.

The required changes to current protocols are minimal and are only in fact
required to run a simulation with the protocol as an entrypoint.

Closes #2233