feat(protocol-designer): add substep level liquid tracking and expose in tooltip (#2400)
Restructure the return value of compound step generators such that their constituent atomic aspirate and dispense actions are not reduced to one step. With the new substep specific intermediate liquid states reshape the way that the steplist generated StepListItems and its children SubstepRows to rely on mid step liquid state, as opposed to initial liquid state of the entire parent step. Instead of passing down a boolean flag to substep rows for presence or absence of liquid in source and destination wells, give them a map of all well liquids and their individual volumes. Display all of this in a HoverTooltip on IngredPill.

Closes #2286