fix(app): Allow portal to re-check for root element
If the portal modal was used during the first render (e.g. for the analytics opt-in modal), the
modal root would not be present in the DOM (since the root itself is created by the App component),
causing the modal to not appear. This change allows the Portal component to a) cache the root
component instead of getting it from the document on every render and b) retry getting the root
after mounting, when we're sure (read: pretty sure) the root is also actually mounted in the DOM