feat(protocol-designer): separate and execute performance audit of labware components
In the 4 main places we are using, or will use, the SelectablePlate component, there are shared
aspects of the abstraction that weigh down the application in places where the extra information is
not needed or used. This branch begins to separate our implementations by first creating a
SelectableLabware component which is used in the WellSelectionModal and maintains local state. The
LabwareOnDeck is also not rendering with select and hover handlers. The IngredientSelectionModal,
will use SelectableLabware or a similar component once the ingredient form sustains upcoming
changes.

Closes #2285