fix(app): Allow portal to re-check for root element (#2440)
If the portal modal was used during the first render (e.g. for the analytics
opt-in modal), the modal root would not be present in the DOM (since the root
itself is created by the App component), causing the modal to not appear. This
change allows the Portal component to:

- Cache the root instead of getting it from the document on every render
- Retry getting the root after mounting, when we're (pretty) sure the root is
  also actually mounted in the DOM