build(api): Add shared-data to wheel
And do it in a way that doesn’t require copying it into the source tree.

By overriding the build_py subcommand of distutils, which is a core element of
basically all the following distutils commands like sdist or install or (most
crucially) bdist_wheel, we can insert our shared-data into the build directory
which means it will be present in all the following artifacts.

Closes #2452