fix(api): Write .env with absolute path for conf on make install
The old checked-in .env had a relative path into tests/ for the development
index.json file that drives settings. This makes it impossible to import
opentrons from outside the api directory. One way to fix this would be to add correct, sane, and not duplicated
logic to the way we load settings; another is to write the .env and the
index.json with absolute paths when you run make install.

Also, this fixes an issue where ‘make local-install’ worked on appveyor (because
appveyor uses mingw make) but doesn’t work on other dev setups that install make
using e.g. chocolatey into the windows shell or powershell, which do not do
wildcard expansion.

Also, a mkdir in discovery-client needs to be shx

Closes #2495