fix(api): Make the Makefile have more real prerequisites/targets
Instead of using all phony targets, the makefile now pulls the version out of
package.json so it can find the name of the wheel that will be built and can use
it as a normal Makefile target. Then, we can use that as the prereq for wheel
instead of having wheel contain the commands directly, which means that all the
stuff that now depends on wheel won’t rerun it every single time it’s called.