feat(protocol-designer): add tooltip for labware name/type on steplist (#2497)
Because of the cramped space on the step list, it is often the case that the labware name is
truncated. For enhanced context, this adds a tooltip that launches on hover of the labware name and
displays the full name and labware type. Also create interface for tooltips to render into a given portal.

Closes #2421