feat(protocol-designer): liquid tooltips on well selection, popper and portal (#2521)
Add functionality to allow users to view contents of wells on hover via a tooltip while selecting
wells to dispense or aspirate from/to in the step forms. To allow reliable positioning of tooltips
in the future, these targets and references are placed inside portals that render them at the page
level, unbound by any intermediate DOM element and the possibility of overly cascading styles. We
can now mend the edge case rendering of large tooltips that exceed the viewport by tapping into the
geometry calculations that Popper uses to flip and reposition if close to boundaries.

Closes #2487