refactor(api): Add instrument load, aspirate, dispense to protocol_api
This has a lot of backend work to prepare the protocol API for actually
interacting with hardware.

We can now load instruments; we can use the backwards-compatibility layer to do
it; and we can call move, aspirate, and dispense on them.

In addition, there is now (autogenerated) documentation for the protocol API. It
is not included in the doctree but can be accessed at
${docsroot}/new_protocol_api.html once built.

Closes #2243