refactor(api): Add a fixture to test with both new and old apis
By using the new old_and_new_client fixture instead of either async_client or
test_client and building a server instance directly, tests can now check both
the old and new versions of the API.

old_and_new_client is a parameterized fixture that pytest will automatically run
once with each parameter, resetting the globals to the correct version.

In addition, the globals in opentrons/__init__ are now defined such that they
can be reset to a different API version by a function call. This has to be used
quite carefully since it will only reset the global instances themselves, not
copies interned elsewhere, but it’s quite useful for this testing purpose.