refactor(api): Make servers use injected hw and new api if ff set
The RPC and HTTP servers (to a large but not complete extent) and the system as
a whole now rely on an injected hardware instance for talking to hardware.

This does not extend to deck calibration, protocol sessions (including labware
calibration), or testing tools.

In addition, there’s work done on the tests so that tests can more easily run on
both versions of the api while relying on simple fixtures; async_client and
main_router and the things that depend on them, for instance, now will run on
both apis unless marked otherwise.

Many small additions have been made to both the new and old APIs to get them to
work a bit more like each other to make the logic necessary in dependent systems
like the servers to be a little more similar between versions.