refactor(api): more functions use hc (#2675)
* refactor(api): Move smoothie update into smoothie driver

This makes it easier to use hardware control instead of the robot in main and is
where the functionality should have been in the first place.

Note that this doesn’t currently apply to the update server.

* refactor(api): Add a fixture to test with both new and old apis

By using the new old_and_new_client fixture instead of either async_client or
test_client and building a server instance directly, tests can now check both
the old and new versions of the API.

old_and_new_client is a parameterized fixture that pytest will automatically run
once with each parameter, resetting the globals to the correct version.

In addition, the globals in opentrons/__init__ are now defined such that they
can be reset to a different API version by a function call. This has to be used
quite carefully since it will only reset the global instances themselves, not
copies interned elsewhere, but it’s quite useful for this testing purpose.

* refactor(api): Change modules to work with control endpoints

We needed the old and new modules to look a bit more alike, and now can get
attached modules and module info from the run app while using hardware_control.

* refactor(api): Pull light control fully into gpio driver

Previously it lived an uncomfortable life half in robot and half in gpio. Since
lights aren’t controlled by smoothie, robot is a bad place for them (but the
calls remain there for back compatibility).

* refactor(api): Make servers use injected hw and new api if ff set

The RPC and HTTP servers (to a large but not complete extent) and the system as
a whole now rely on an injected hardware instance for talking to hardware.

This does not extend to deck calibration, protocol sessions (including labware
calibration), or testing tools.

In addition, there’s work done on the tests so that tests can more easily run on
both versions of the api while relying on simple fixtures; async_client and
main_router and the things that depend on them, for instance, now will run on
both apis unless marked otherwise.

Many small additions have been made to both the new and old APIs to get them to
work a bit more like each other to make the logic necessary in dependent systems
like the servers to be a little more similar between versions.

* refactor(api): add ability to override critical point in moves and get position

This is needed for the move endpoint because it makes it a lot easier to move
pipettes to a change pipette position guaranteed to neither be too high nor too low.

Closes #2245