refactor(api): Use modules in protocol api (#2702)
* refactor(api): Use modules in protocol api

There are now ModuleContexts in protocol_api/contexts.py that expose the module
API. The protocol API also has a new workflow for adding labware to modules:

- create a module with ProtocolContext.load_module()
- call module.load_labware[_by_name]() to load the labware

This avoids the shared=True functionality that was only used for modules but was
designed to be super generic.

Modules also now have their own specs file that defines their geometry,
including the offsets from the deck to their labware and their overall shape.

In addition, fix some minor bugs:
- axis maxima bumped by +0.05mm for all axes to avoid floating point precision
problems when positions are run through deck transform
- hardware_control SynchronousAdapter needs to actually return the results of
coroutines it executes
- Reset the protocol context location cache after homing

feature(api): Add magnetic module compatiblity to labware schema

Labware now has an extra couple of parameters. isMagneticModuleCompatible is a
bool indicating whether the labware can be easily used by a magdeck;
magneticModuleEngageHeight is the height to raise the magnets for that piece of
labware. If a labware that is not compatible is loaded onto a magdeck, you get a
warning and must explicitly specify the magnet height.

* Fix failing tests