refactor(step-generation): create getNextRobotStateAndWarnings and migrate aspirate
In an effort to consolidate the logic in step generation that calculates the next robotState given a
command, this adds a function called getNextRobotStateAndWarnings. Eventually this will be used in
all command creators to generate the next robotState given StepArgs. This takes the first step by
migrating the aspirate commandCreator from aspirateUpdateLiquidState to the new
getNextRobotStateAndWarnings

Closes #2734