refactor(api): Make SynchAdapter run its event loop in a thread
This is important to sidestep issues with pytest hijacking event loops or
different things calling coroutines at different times. Now, the event loop is
canonically run by the synchadapter thread and calls are submitted to it with
run_coroutine_threadsafe by all callers.

This comes with a bit of extra work in terms of lifecycle management for the
synchadapter, which is where the changes to the context come in. Using the
hardware manager rather than the hardware directly reduces the number of places
that have direct refs to the hardware and now there’s some logic that lets us
kill the unused contexts.