fix(api): Do not overwrite settings every time we get config files
opentrons.config._generate_base_config(), in addition to generating and
returning defaults, will _write_ those defaults to the appropriate places unless
the skip_usb flag is set. In get_config_index(), which is mostly intended to
just find the config files, it uses generate_base_config to generate the config
but does not intend to write it. We therefore need to explicitly tell
_generate_base_config to not overwrite anything.