feat(api): Add ability to use papi2 in protocol (#2803)
* refactor(api): Do not expose an "axis" member of the pipette model

We shouldn’t be exposing an axis here; and this isn’t even the actual axis,
since it can be either "a" or "b", which would be the right pipette Z axis and
the left pipette plunger axis respectively.

* refactor(api): Add broker logging to new protocol api

* refactor(api): Run protocols in the new protocol API

Protocols should have a run(context) function and should call methods on the
context to actually do actions in their protocols.

* refactor(api): Make SynchAdapter run its event loop in a thread

This is important to sidestep issues with pytest hijacking event loops or
different things calling coroutines at different times. Now, the event loop is
canonically run by the synchadapter thread and calls are submitted to it with
run_coroutine_threadsafe by all callers.

This comes with a bit of extra work in terms of lifecycle management for the
synchadapter, which is where the changes to the context come in. Using the
hardware manager rather than the hardware directly reduces the number of places
that have direct refs to the hardware and now there’s some logic that lets us
kill the unused contexts.

* refactor(api): Use new protocol api in sessions

* refactor(api): Out of bounds is now a warning rather than an err

It’s very annoying.

* feat(api): Add protocol error checking in a better way

New protocols now raise a special error interpreted by the RPC differently than
other exceptions. This allows the new protocol execute handler to splice in the
correct line number information.

* fixup(api): Fix some errors noticed by lint and add some errors

* fixup(api): Change annotations and errors for input to some functions

drop_tip and load_instrument and pick_up_tip now have a couple more options for
how to get input - specifying wells for drop_tip and pick_up_tip and a string
for the mount name in load_instrument. In addition, reworked docstrings and
error messages.

* fixup(api): Get proper labware reporting to runapp

* feat(api): Add configurations for the tip overlap distance

Tips overlap with the pipette nozzles by a distance that, while varying with the
specific pipette, tip brand, tip, and pick up settings does have a single
nominal value for a kind of pipette. Those are not all yet encoded, but this is
a start.

* fixup(api): Fix unit tests for tip overlap calculation

* fixup(api): Add high level docs for new protocol API

* fixup(api): Add depth to tiprack wells and drop tips to bottom()

* fixup(api): Fix new test after rebase