feat(api): api2: Add translation between old and new labwares
This adds a translation table so (if the back compat layer is used, and in json
protocols) if a labware is loaded by old name and we have a new definition for
it, we’ll just patch in the new definition. If it’s an old name and we don’t
have a new definition, we also have a nice error.