feat(api): Add json protocol execution to new protocol API (#2854)
Add the ability to execute json protocols to the new protocol API.

Add a translation table so (if the back compat layer is used, and in json
protocols) if a labware is loaded by old name and we have a new definition for
it, we’ll just patch in the new definition. If it’s an old name and we don’t
have a new definition, we also have a nice error.

Do not raise exceptions on mismatched pipette or module state; instead, trust the frontend to present these issues to the user.

Add the tip overlap to the schema. Load it and fold it into tiprack tip length
definition at runtime. This lets us keep the actual nominal tip length where it
should be in the defs while taking into account the overlap.

Closes #2248.